remove unneeded deploys, clean
