+++ scripts/deploy/deploy-rUSDC-ETH-P-THETA-stakingrewards.js
